chore: more directory changes
