+++ apps/dashboard/app/(layowt)/(dashboard)/dashboard/[uid]/page-client.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { setWebsite } from '@/utils/websites/setWebsite';
+import { setWebsite } from '@/actions/websites/set-website';
 import { websites } from '@prisma/client';
 
 export default function PageClient({ website }: { website: websites }) {
