+++ apps/dashboard/app/(layowt)/(dashboard)/dashboard/page.tsx
@@ -1,8 +1,7 @@
 'use server';
 import WebsiteCard from '@/components/layout/dashboard/website-card';
-import { getUserFromSession } from '@/utils/user';
-import { getUserFromDb } from '@/utils/user';
-import { getWebsite } from '@/utils/websites';
+import { getUserFromSession, getUserFromDb } from '@/actions/user/get-user';
+import { getWebsite } from '@/actions/websites/get-website';
 import { websites } from '@prisma/client';
 import PageClient from './page-client';
