+++ apps/dashboard/components/auth/login-form.tsx
@@ -11,13 +11,11 @@ import { Input } from '@/ui/input';
 import { useState } from 'react';
 import { Button } from '@/ui/button';
 import { toast } from 'sonner';
-import { login } from '@/utils/user/';
+import { login } from '@/actions/user/user-login';
 import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { useSearchParams } from 'next/navigation'
-import { useUser } from '@/hooks/useUser';
-import { prisma } from '@/utils/prisma';
-import { getWebsite } from '@/utils/websites';
+import { getWebsite } from '@/actions/websites/get-website';
 
 export default function LoginForm() {
   const router = useRouter();
