+++ apps/dashboard/components/layout/dashboard/navigation/sidebar.tsx
@@ -1,6 +1,6 @@
 'use server';
-import { getUserFromSession } from '@/utils/user';
-import { getWebsite } from '@/utils/websites';
+import { getUserFromSession } from '@/actions/user/get-user';
+import { getWebsite } from '@/actions/websites/get-website';
 import type { websites } from '@prisma/client';
 
 import NavigationItems from '@/components/layout/dashboard/navigation/items';
