+++ apps/dashboard/components/layout/site-builder/navigation/publish-dropdown-items.tsx
@@ -7,7 +7,8 @@ import { DropdownMenuSeparator } from '@/components/ui/dropdown-menu';
 
 // utils
 import type { websites as Website } from '@prisma/client';
-import { publishSite, updateWebsite } from '@/utils/websites';
+import { updateWebsite } from '@/actions/websites/update';
+import { publishSite } from '@/actions/websites/publish-website';
 import { getTimeStamp } from '@/utils/index';
 import { User } from '@supabase/supabase-js';
 import { useUser } from '@/hooks/useUser';
